+++ spec/unit/outputs/cassandra_spec.rb
@@ -1,3 +1,3 @@
 # encoding: utf-8
-require_relative "../cassandra_spec_helper"
+require_relative "../../cassandra_spec_helper"
 require "logstash/outputs/cassandra_output"
