+++ spec/unit/outputs/event_parser_spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require_relative "../cassandra_spec_helper"
+require_relative "../../cassandra_spec_helper"
 require "logstash/outputs/cassandra/event_parser"
 
 RSpec.describe LogStash::Outputs::Cassandra::EventParser do
